# msbuild: Merge "Core" libs into single DolphinLib