--- a/Source/Core/Common/Common.vcxproj
+++ b/Source/Core/Common/Common.vcxproj
@@ -53,7 +53,6 @@
     <ClInclude Include="DynamicLibrary.h" />
     <ClInclude Include="ENetUtil.h" />
     <ClInclude Include="Event.h" />
-    <ClInclude Include="File.h" />
     <ClInclude Include="FileSearch.h" />
     <ClInclude Include="FileUtil.h" />
     <ClInclude Include="FixedSizeQueue.h" />
@@ -124,6 +123,7 @@
     <ClInclude Include="HttpRequest.h" />
     <ClInclude Include="Image.h" />
     <ClInclude Include="IniFile.h" />
+    <ClInclude Include="IOFile.h" />
     <ClInclude Include="JitRegister.h" />
     <ClInclude Include="Lazy.h" />
     <ClInclude Include="LdrWatcher.h" />
@@ -194,7 +194,6 @@
     <ClCompile Include="Debug\Watches.cpp" />
     <ClCompile Include="DynamicLibrary.cpp" />
     <ClCompile Include="ENetUtil.cpp" />
-    <ClCompile Include="File.cpp" />
     <ClCompile Include="FileSearch.cpp" />
     <ClCompile Include="FileUtil.cpp" />
     <ClCompile Include="FloatUtils.cpp" />
@@ -216,6 +215,7 @@
     <ClCompile Include="HttpRequest.cpp" />
     <ClCompile Include="Image.cpp" />
     <ClCompile Include="IniFile.cpp" />
+    <ClCompile Include="IOFile.cpp" />
     <ClCompile Include="JitRegister.cpp" />
     <ClCompile Include="LdrWatcher.cpp" />
     <ClCompile Include="Logging\ConsoleListenerWin.cpp" />