--- a/Source/Core/Core/DolphinAnalytics.cpp
+++ b/Source/Core/Core/DolphinAnalytics.cpp
@@ -1,4 +1,4 @@
-#include "Core/Analytics.h"
+#include "Core/DolphinAnalytics.h"
 
 #include <array>
 #include <memory>