--- a/Source/Core/DiscIO/WiiSaveBanner.cpp
+++ b/Source/Core/DiscIO/WiiSaveBanner.cpp
@@ -10,7 +10,7 @@
 
 #include "Common/ColorUtil.h"
 #include "Common/CommonTypes.h"
-#include "Common/File.h"
+#include "Common/IOFile.h"
 #include "Common/NandPaths.h"
 #include "Common/StringUtil.h"