… Externals/libusb/libusb-1.0.vcxproj
… Source/Core/AudioCommon/AudioCommon.vcxproj
… Source/Core/AudioCommon/AudioCommon.vcxproj.filters
… Source/Core/Common/CMakeLists.txt
… Source/Core/Common/IOFile.h
… Source/Core/Core/Boot/Boot.cpp
… Source/Core/Core/CMakeLists.txt
… Source/Core/Core/ConfigManager.cpp
… Source/Core/Core/DolphinAnalytics.h
… Source/Core/Core/HW/EXI/EXI_DeviceIPL.cpp
… Source/Core/Core/HW/GCMemcard/GCMemcardDirectory.cpp
… Source/Core/Core/HW/GCMemcard/GCMemcardRaw.cpp
… Source/Core/Core/HW/WiiSave.cpp
… Source/Core/Core/HW/WiimoteEmu/EmuSubroutines.cpp
… Source/Core/Core/IOS/DI/DI.cpp
… Source/Core/Core/IOS/FS/FileSystemCommon.cpp
… Source/Core/Core/IOS/IOSC.cpp
… Source/Core/Core/Movie.cpp
… Source/Core/Core/NetPlayClient.cpp
… Source/Core/Core/State.cpp
… Source/Core/DiscIO/DirectoryBlob.cpp
… Source/Core/DiscIO/DiscIO.vcxproj
… Source/Core/DiscIO/DiscIO.vcxproj.filters
… Source/Core/DiscIO/VolumeVerifier.cpp
… Source/Core/DiscIO/WIABlob.cpp
… Source/Core/DolphinLib.ARM64.props
… Source/Core/DolphinLib.props
… Source/Core/DolphinLib.vcxproj
… Source/Core/DolphinLib.x64.props
… Source/Core/DolphinNoGUI/DolphinNoGUI.vcxproj