@Rumi-Larry 2020-09-20
It might be a good idea to rename "VulkanContext" to "VKContext" for consistency 🤷