@shuffle2 2020-10-27
substantive/constructive comments appreciated ...