@shuffle2 2020-10-27
It was too long ago now, I'll have to redo the perf traces and get back about exact numbers, but it cut full rebuild time for me into about half iirc.

Anyone can look themselves with vcperf.