@shuffle2 2021-01-10
Some simple perf stats of running the following for both current master and this PR:
```
rmdir /s/q Build Binary
msbuild Source\dolphin-emu.sln -p:Configuration=Release,Platform=x64 -m -v:q -clp:PerformanceSummary
```
master:
```
Task Performance Summary:
        1 ms  ConvertToAbsolutePath                     47 calls
        3 ms  RemoveDuplicates                          94 calls
        7 ms  AssignCulture                             47 calls
        8 ms  CreateProperty                           244 calls
        9 ms  HostTranslatePaths                       122 calls
        9 ms  CreateItem                                94 calls
       13 ms  Message                                  173 calls
       15 ms  AssignTargetPath                         329 calls
       16 ms  FindUnderPath                            235 calls
       16 ms  CheckVCToolsetVersion                     47 calls
       21 ms  GetItemHash                              122 calls
       28 ms  AssignProjectConfiguration                47 calls
       29 ms  Flatten                                  122 calls
       42 ms  Expand                                     1 calls
       53 ms  SetEnv                                   376 calls
       55 ms  ReadLinesFromFile                         95 calls
       62 ms  GetOutOfDateItems                        141 calls
       66 ms  Delete                                    94 calls
       73 ms  GenerateDesktopDeployRecipe               47 calls
       83 ms  WriteLinesToFile                         203 calls
      121 ms  RC                                         2 calls
      181 ms  Touch                                     94 calls
      260 ms  QtRunWork                                  1 calls
      330 ms  MakeDir                                  582 calls
      541 ms  CallTarget                                98 calls
      884 ms  Copy                                     2003 calls
     3452 ms  Link                                       4 calls
     3508 ms  Exec                                       1 calls
     8854 ms  msgfmt                                    28 calls
    18039 ms  LIB                                       40 calls
    262487 ms  CL                                        47 calls
    511553 ms  MSBuild                                  132 calls
```
PR:
```
Task Performance Summary:
        4 ms  CreateProperty                           244 calls
        4 ms  ConvertToAbsolutePath                     33 calls
        4 ms  CreateItem                                66 calls
        5 ms  AssignCulture                             33 calls
        7 ms  RemoveDuplicates                          66 calls
       10 ms  HostTranslatePaths                       122 calls
       10 ms  CheckVCToolsetVersion                     33 calls
       12 ms  Message                                  131 calls
       15 ms  FindUnderPath                            165 calls
       20 ms  AssignProjectConfiguration                33 calls
       25 ms  SetEnv                                   264 calls
       26 ms  GetItemHash                              122 calls
       32 ms  AssignTargetPath                         231 calls
       37 ms  ReadLinesFromFile                         67 calls
       39 ms  Flatten                                  122 calls
       41 ms  Expand                                     1 calls
       44 ms  GetOutOfDateItems                         99 calls
       80 ms  GenerateDesktopDeployRecipe               33 calls
      100 ms  WriteLinesToFile                         189 calls
      104 ms  Delete                                    66 calls
      220 ms  CallTarget                                70 calls
      237 ms  RC                                         2 calls
      248 ms  Touch                                     66 calls
      261 ms  QtRunWork                                  1 calls
      462 ms  MakeDir                                  442 calls
      907 ms  Copy                                     2003 calls
     3595 ms  Link                                       4 calls
     7133 ms  Exec                                       1 calls
    10613 ms  msgfmt                                    28 calls
    23056 ms  LIB                                       26 calls
    169739 ms  MSBuild                                   54 calls
    253276 ms  CL                                        33 calls
```